+++ src/main/kotlin/net/alexanderkahn/longball/service/model/RosterPlayer.kt
@@ -2,4 +2,4 @@ package net.alexanderkahn.longball.service.model
 
 import java.time.ZonedDateTime
 
-data class RosterPlayer (val id: Long, val team: Long, val player: Long, val jerseyNumber: Short, val startDate: ZonedDateTime, val endDate: ZonedDateTime? = null)
+data class RosterPlayer (val id: Long, val team: Long, val player: Long, val jerseyNumber: Int, val startDate: ZonedDateTime, val endDate: ZonedDateTime? = null)
